refactor: Common/user/business 디렉토리 내 파일 이동
- 구현체 외부에 존재 : Common/user/business/external
- common service, entity service : Common/user/business/internal
