+++ app/src/test/java/com/food/api/AuthenticationTest.java
@@ -4,8 +4,8 @@ import com.food.SuperIntegrationTest;
 import com.food.auth.filter.dto.AuthenticatedUser;
 import com.food.auth.provider.AccessTokenProvider;
 import com.food.auth.provider.dto.AccessToken;
-import com.food.common.user.business.external.dto.AppAccountDto;
-import com.food.common.user.business.service.response.accountFind.AccountFindResponse;
+import com.food.common.user.business.internal.dto.AppAccountDto;
+import com.food.common.user.business.external.model.AccountFindResponse;
 import com.food.common.user.domain.AppAccount;
 import com.food.common.user.domain.User;
 import com.food.common.user.enumeration.Role;
