+++ app/src/test/java/com/food/api/auth/TokenRenewApiTest.java
@@ -1,8 +1,8 @@
 package com.food.api.auth;
 
 import com.food.SuperIntegrationTest;
-import com.food.common.user.business.external.RefreshTokenCommonService;
-import com.food.common.user.business.external.dto.RefreshTokenDto;
+import com.food.common.user.business.internal.RefreshTokenCommonService;
+import com.food.common.user.business.internal.dto.RefreshTokenDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
