+++ auth/src/main/java/com/food/auth/business/DefaultAuthenticatedUserFindService.java
@@ -1,7 +1,7 @@
 package com.food.auth.business;
 
-import com.food.common.user.business.service.AuthenticatedUserFindService;
-import com.food.common.user.dto.RequestUser;
+import com.food.common.user.business.external.AuthenticatedUserFindService;
+import com.food.common.user.business.external.model.RequestUser;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
