+++ auth/src/main/java/com/food/auth/business/LoginService.java
@@ -5,14 +5,14 @@ import com.food.auth.presentation.dto.LoginRequest;
 import com.food.auth.presentation.dto.TokenIssueResponse;
 import com.food.auth.presentation.dto.TokenRenewResponse;
 import com.food.auth.provider.AccessTokenProvider;
-import com.food.common.user.business.external.AccountCommonService;
-import com.food.common.user.business.external.LogInOutLogCommonService;
-import com.food.common.user.business.external.RefreshTokenCommonService;
-import com.food.common.user.business.external.dto.AppAccountDto;
-import com.food.common.user.business.external.dto.LogInOutLogRequest;
-import com.food.common.user.business.external.dto.RefreshTokenDto;
-import com.food.common.user.business.external.dto.SocialAccountDto;
-import com.food.common.user.dto.RequestUser;
+import com.food.common.user.business.internal.AccountCommonService;
+import com.food.common.user.business.internal.LogInOutLogCommonService;
+import com.food.common.user.business.internal.RefreshTokenCommonService;
+import com.food.common.user.business.internal.dto.AppAccountDto;
+import com.food.common.user.business.internal.dto.LogInOutLogRequest;
+import com.food.common.user.business.internal.dto.RefreshTokenDto;
+import com.food.common.user.business.internal.dto.SocialAccountDto;
+import com.food.common.user.business.external.model.RequestUser;
 import com.food.common.user.enumeration.AccountType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
