+++ auth/src/main/java/com/food/auth/filter/AccessTokenAuthenticationFilter.java
@@ -5,8 +5,8 @@ import com.food.auth.provider.AccessTokenProvider;
 import com.food.auth.provider.dto.AccessToken;
 import com.food.auth.provider.dto.AccessTokenContent;
 import com.food.auth.provider.dto.AccessTokenValidationResult;
-import com.food.common.user.business.service.AccountFindService;
-import com.food.common.user.business.service.response.accountFind.AccountFindResponse;
+import com.food.common.user.business.external.AccountFindService;
+import com.food.common.user.business.external.model.AccountFindResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
