+++ common/src/main/java/com/food/common/user/business/external/AccountFindService.java
@@ -1,6 +1,6 @@
-package com.food.common.user.business.service;
+package com.food.common.user.business.external;
 
-import com.food.common.user.business.service.response.accountFind.AccountFindResponse;
+import com.food.common.user.business.external.model.AccountFindResponse;
 
 import javax.validation.constraints.NotNull;
