+++ common/src/main/java/com/food/common/user/business/external/AuthenticatedUserFindService.java
@@ -1,0 +1,7 @@
+package com.food.common.user.business.external;
+
+import com.food.common.user.business.external.model.RequestUser;
+
+public interface AuthenticatedUserFindService {
+    RequestUser findRequestUser();
+}
