--- common/src/main/java/com/food/common/user/business/service/AuthenticatedUserFindService.java
@@ -1,7 +1,0 @@
-package com.food.common.user.business.service;
-
-import com.food.common.user.dto.RequestUser;
-
-public interface AuthenticatedUserFindService {
-    RequestUser findRequestUser();
-}
